requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.3
scipy==1.15.2
threadpoolctl==3.5.0
regex==2024.11.6
googletrans==4.0.0-rc1
